train-uploader/requirements.txt
flask
python-dotenv
GitPython
